Replace Hatch with Poetry and add Python Test in Github Workflows
In this commit, the switch from Hatch to Poetry is made in the project's setup documents. A Python Test workflow is added to Github Actions to automate Python testing. The project's setup documents are updated to better represent the switch to Poetry. The license file was also deleted for unspecified reasons.